refactor: move removing logic to the FloorPictureList component
